VC++ 2008 프로젝트에 x64 플랫폼을 추가함. x64 플랫폼에서 빌드 경고 메시지가 출력되지 않도록 수정함.

+++ trunk/SipParser/SipChallenge.cpp
@@ -165,7 +165,7 @@ void CSipChallenge::Clear()
 
 void CSipChallenge::DeQuoteString( std::string & strInput, std::string & strOutput )
 {
-	int iLen = strInput.length();
+	int iLen = (int)strInput.length();
 
 	strOutput.clear();
